Move syncing logic out of Editor into hook
