+++ frontend/src/hooks/index.js
@@ -8,3 +8,6 @@ export { default as useTransitionCreation } from './useTransitionCreation'
 export { default as useStateSelection } from './useStateSelection'
 export { default as useTransitionSelection } from './useTransitionSelection'
 export { default as useContextMenus } from './useContextMenus'
+export { default as useSyncProjects } from './useSyncProjects'
+export { default as useSyncCurrentProject } from './useSyncCurrentProject'
+export { default as useAutosaveProject } from './useAutosaveProject'
