+++ frontend/src/index.js
@@ -1,14 +1,12 @@
 import { StrictMode, createElement } from 'react'
 import ReactDOM from 'react-dom'
 import { setup } from 'goober'
-import { BrowserRouter, Link, Route, Routes, useLocation } from 'react-router-dom'
+import { BrowserRouter, Route, Routes, useLocation } from 'react-router-dom'
 
 import * as Pages from './pages'
 
-import { useEgg } from '/src/hooks'
+import { useEgg, useSyncProjects } from '/src/hooks'
 import { Footer } from '/src/components'
-import useSyncProjects from '/src/hooks/useSyncProjects'
-import useAutosaveProject from './hooks/useAutosaveProject'
 
 // Set up goober to use React
 setup(
